Add files via upload
Should be easy to integrate in i2c system as slave.
NOTE: If this is ran without a physical stop on the far left of the rail: during startup it will go left until something breaks.
The intent is to have a stop/object nudge the arm when it hits zero position. Necessary for relative movement.